+++ include/pinocchio/src/constraints/utils.hxx
@@ -942,8 +942,13 @@ namespace pinocchio
         PINOCCHIO_UNUSED_VARIABLE(block_infos);
         PINOCCHIO_UNUSED_VARIABLE(dispatcher);
 
+        // For some reason this assert is always evaluated when building
+        // with g++ 11.
+        // TODO Remove when Ubuntu 22.04 is no more supported.
+#if !defined(__GNUC__) || __GNUC__ >= 12
         static_assert(
           false, "ComputeBlockDiagonalPatternImpl not implemented for this constraint.");
+#endif // !defined(__GNUC__) || __GNUC__ >= 12
       }
     };